lookAt and ortho actually working (but actually this time)

+++ src/Graphics/GMesh.h
@@ -17,6 +17,6 @@ private:
 	int drawMode;
 	int drawCount;
 	int drawComponentType;
-	int drawByteOffset;
+	size_t drawByteOffset;
 };
